Expense changes

+++ db/migrate/20170213185329_create_expenses.rb
@@ -2,7 +2,7 @@ class CreateExpenses < ActiveRecord::Migration[5.0]
   def change
     create_table :expenses do |t|
       t.string :category
-      t.integer :budget
+      t.float :budget
       t.integer :user_id
 
       t.timestamps
